deploy_ai/API/requirements_freeze.txt
absl-py==0.7.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
astpretty==1.4.0
beautifulsoup4==4.6.3
bleach==1.5.0
blinker==1.4
boto==2.49.0
boto3==1.9.28
botocore==1.12.28
Brlapi==0.6.6
bz2file==0.98
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
checkbox-support==0.22
Click==7.0
cloudpickle==0.7.0
command-not-found==0.3
cryptography==2.4.2
cupshelpers==1.0
cycler==0.10.0
dask==1.1.1
deap==1.2.2
decorator==4.3.2
defer==1.0.6
distro-info==0.18
docopt==0.6.2
docutils==0.14
feedparser==5.2.1
fire==0.1.3
Flask==1.0.2
gast==0.2.2
gensim==3.6.0
grpcio==1.19.0
guacamole==0.9.2
h5py==2.9.0
html5lib==0.9999999
httplib2==0.9.2
idna==2.8
imageio==2.4.1
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lmdb==0.94
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mock==2.0.0
moviepy==0.2.3.5
netifaces==0.10.4
networkx==2.2
nltk==3.4
numpy==1.15.4
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
onboard==1.4.1
opencv-python==4.0.0.21
padme==1.1.1
pbr==5.1.3
pexpect==4.2.1
Pillow==5.4.1
pipreqs==0.4.9
plainbox==0.25
protobuf==3.7.0
pyasn1==0.4.2
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.3.1
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.7.5
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.0.1
pyxdg==0.25
PyYAML==3.12
regex==2017.4.5
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
s3transfer==0.1.13
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.7.1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.8.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
textblob==0.15.2
toolz==0.9.0
tqdm==4.29.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.24.1
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
webencodings==0.5
Werkzeug==0.14.1
xdiagnose==3.8.8
xkit==0.0.0
XlsxWriter==0.9.6
yarg==0.1.9
zope.interface==4.3.2
